fix: applicationOrder 를 int 자료형으로 처리하도록 복구
- 순서를 나타내고, 해당 값이 21억을 넘길 수 없다 판단하여 더 적합한 int 자료형으로 복구

* test: long type 을 기대하던 테스트 에러 해결

+++ src/main/java/com/example/solidconnection/admin/dto/MentorApplicationHistoryResponse.java
@@ -8,7 +8,7 @@ public record MentorApplicationHistoryResponse(
         MentorApplicationStatus mentorApplicationStatus,
         String rejectedReason,
         ZonedDateTime createdAt,
-        long applicationOrder
+        int applicationOrder
 ) {
 
 }
